first fragment created
