+++ app/build.gradle
@@ -68,4 +68,9 @@ dependencies {
     kapt("androidx.room:room-compiler:$room_version")
 
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.9'
+
+    //fragments
+    def fragment_version = "1.3.6"
+    implementation "androidx.fragment:fragment-ktx:$fragment_version"
+
 }
